move unittests of AxesToDataScale in Figure tests
